+++ test/test_models.py
@@ -1,5 +1,6 @@
 import unittest
-from game.models import Tiles, BagTiles, Player, Board, Cell
+from game.models import Tiles, BagTiles, Player, Cell
+from game.board import Board
 from unittest.mock import patch
 
 class TestTiles(unittest.TestCase):
@@ -42,69 +43,11 @@ class TestBagTiles(unittest.TestCase):
             len(bag.tiles), 
             102,)
         
-
 class TestPlayer(unittest.TestCase):
     def test_init(self):
         bag_tiles = BagTiles()
         player_1 = Player(1, bag_tiles)
         self.assertEqual(len(player_1.tiles),7,)
-
-class TestBoard(unittest.TestCase):
-    def test_init(self):
-        board = Board()
-        self.assertEqual(len(board.grid),15,)
-        self.assertEqual(len(board.grid[0]),15,)
-    
-    def test_positions_word_3(self):
-        board = Board()
-        board.positions()
-        self.assertEqual(board.grid[0][0].multiplier, 3)
-        self.assertEqual(board.grid[0][0].multiplier_type, 'word')
-        self.assertEqual(board.grid[7][0].multiplier, 3)
-        self.assertEqual(board.grid[7][0].multiplier_type, 'word')
-        self.assertEqual(board.grid[14][14].multiplier, 3)
-        self.assertEqual(board.grid[14][14].multiplier_type, 'word')
-        self.assertNotEqual(board.grid[0][14], board.grid[5][7])
-
-    def test_positions_word_3(self):
-        board = Board()
-        board.positions()
-        self.assertEqual(board.grid[1][1].multiplier, 2)
-        self.assertEqual(board.grid[2][12].multiplier_type, 'word')
-        self.assertEqual(board.grid[11][3].multiplier, 2)
-        self.assertEqual(board.grid[1][13].multiplier_type, 'word')
-        self.assertNotEqual(board.grid[0][14].multiplier, 2)
-        self.assertNotEqual(board.grid[7][7].multiplier, 2)
-        self.assertNotEqual(board.grid[6][7].multiplier, 2)
-
-    def test_positions_letter_2(self):
-        board = Board()
-        board.positions()
-        self.assertEqual(board.grid[0][3].multiplier, 2)
-        self.assertEqual(board.grid[0][3].multiplier_type, 'letter')
-        self.assertEqual(board.grid[6][6].multiplier, 2)
-        self.assertEqual(board.grid[6][6].multiplier_type, 'letter')
-        self.assertNotEqual(board.grid[0][14].multiplier, 2)
-        self.assertNotEqual(board.grid[7][7].multiplier, 2)
-        self.assertNotEqual(board.grid[6][1].multiplier, 2)
-
-    def test_word_inside_board(self):
-        board = Board()
-        word = "Facultad"
-        location = (5, 4)
-        orientation = "H"
-        word_is_valid = board.validate_word_inside_board(word, location, orientation)
-        assert word_is_valid == True
-        
-    def test_word_out_of_board(self):
-        board = Board()
-        word = "Facultad"
-        location = (14, 4)
-        orientation = "H"
-        word_is_valid = board.validate_word_inside_board(word, location, orientation)
-        
-        assert word_is_valid == False
-
 
 class TestCell(unittest.TestCase):
     def test_init(self):
@@ -127,7 +70,6 @@ class TestCell(unittest.TestCase):
         letter = Tiles(letter = 'P', value = 3)
         cell.add_letter(letter)
         self.assertEqual(cell.calculate_value(),6)
-
 
 class TestCalculateWordValue(unittest.TestCase):
     def test_simple(self):